Moved the USB struct templates
