+++ src/helper/usb_struct_templates.h
